+++ src/Identity/Logitar.Identity.Domain/Sessions/ISessionManager.cs
@@ -1,7 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Sessions;
-
-public interface ISessionManager
-{
-  Task DeleteAsync(SessionAggregate session, CancellationToken cancellationToken = default);
-  Task SaveAsync(SessionAggregate session, CancellationToken cancellationToken = default);
-}