+++ src/Identity/Logitar.Identity.Domain/Settings/ISettingsResolver.cs
@@ -1,9 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Settings;
-
-public interface ISettingsResolver
-{
-  IRoleSettings RoleSettings { get; }
-  IUserSettings UserSettings { get; }
-
-  IPbkdf2Settings Pbkdf2Settings { get; }
-}