+++ src/Identity/Logitar.Identity.Domain/Users/IEmailAddress.cs
@@ -1,6 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Users;
-
-public interface IEmailAddress
-{
-  string Address { get; }
-}