+++ src/Identity/Logitar.Identity.EntityFrameworkCore.Relational/Entities/UserRoleEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace Logitar.Identity.EntityFrameworkCore.Relational.Entities;
-
-public record UserRoleEntity
-{
-  public int UserId { get; set; }
-  public int RoleId { get; set; }
-}