+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Marked old TypeExtensions as obsolete.
 
+### Removed
+
+- Removed Identity projects.
+
 ## [1.7.0] - 2023-09-21
 
 ### Added